requirements.txt
streamlit
opencv-python
numpy
torch
onnx
onnxruntime
matplotlib
scikit-learn
ultralytics
Pillow
mindspore
tqdm